Add OpenMP progress tracing using OMPT (#1)
As a first draft, we will keep MPI and OMP profiling independent. This
is not necessarily a great idea, since we should expect most
applications to be MPI+OpenMP, but until we figure out how each
instrumentation works, it easier that way.

Note that this result in a different downstream publisher than the PMPI
interface.

Code generator included as an extra.